macports: non-sudo build
Improves separation of builds on Arm/M1 and simplifies the build.

Save build logs on fail

+++ scripts/macports_uninstall.sh
@@ -22,12 +22,18 @@
  # Boston, MA  02110-1301,  USA       gnu@gnu.org                   #
  ####################################################################
 
-source ~/.profile
-export PATH=$PREFIX/bin:$PATH
-
+if [[ $(uname -m) == 'arm64' ]]; then
+  build_arm64=true
+  arch='arm64'
+  echo "*** Build: arm64"
+else
+  build_arm64=false
+  arch='x86_64'
+  echo "*** Build: x86_64"
+fi
 
 function pure_version() {
-	echo '0.1'
+	echo '0.2'
 }
 
 function version() {
@@ -44,6 +50,10 @@ function usage() {
     echo "  --formulas-only"
     echo "      only uninstall formulas, not macports itself. Needed for CI where deleting users requires"
     echo "      accepting a dialog box."
+    echo "  --homedirgimp2"
+    echo "      installs macports to a custom homedir macports-gimp2"
+    echo "  --homedirgimp3"
+    echo "      installs macports to a custom homedir macports-gimp3"
     echo "  --version         show tool version number"
     echo "  -h, --help        display this help"
     exit 0
@@ -56,6 +66,12 @@ while test "${1:0:1}" = "-"; do
 	--formulas-only)
 		formulasonly="true"
 		shift;;
+	--homedirgimp2)
+		homedirgimp2="true"
+		shift;;
+	--homedirgimp3)
+		homedirgimp3="true"
+		shift;;
 	-h | --help)
 		usage;;
 	--version)
@@ -66,13 +82,36 @@ while test "${1:0:1}" = "-"; do
 	esac
 done
 
+if [ -n "$homedirgimp2" ]; then
+  echo "**Uninstalling MacPorts in home dir macports-gimp2"
+  home_dir=true
+  PREFIX="${HOME}/macports-gimp2-${arch}"
+  unset dosudo
+elif [ -n "$homedirgimp3" ]; then
+  echo "**Uninstalling MacPorts in home dir macports-gimp3"
+  home_dir=true
+  PREFIX="${HOME}/macports-gimp3-${arch}"
+  unset dosudo
+else
+  PREFIX=/opt/local
+  dosudo=sudo
+fi
+
+export PATH=$PREFIX/bin:$PATH
+
+echo "Macports installed at $(which port &> /dev/null)"
 if which port &> /dev/null; then
   echo "Uninstalling formulas"
   $dosudo port -fp uninstall installed || true
-  # Left behind when uninstalling
-  $dosudo rm -rf /Library/LaunchDaemons/org.macports.rsyncd.plist
-  $dosudo rm -rf /Library/LaunchAgents/org.freedesktop.dbus-session.plist
-  $dosudo rm -rf /Library/LaunchDaemons/org.freedesktop.dbus-system.plist
+  # Following deletes can be removed once not installing startup items has
+  # been rolled out to all builds
+  if [ -n "${dosudo}" ]; then
+    # Due to nightly
+    $dosudo rm -rf /Library/LaunchDaemons/org.macports.rsyncd.plist
+    $dosudo rm -rf /Library/LaunchAgents/org.freedesktop.dbus-session.plist
+    $dosudo rm -rf /Library/LaunchDaemons/org.freedesktop.dbus-system.plist
+    $dosudo rm -rf /Library/LaunchDaemons/org.macports.shared-mime-info-updater.plist
+  fi
 fi
 
 if [ -n "$formulasonly" ]; then
@@ -81,7 +120,7 @@ fi
 
 echo "Uninstalling MacPorts"
 if [ -z "${dosudo}" ]; then
-  if [ -n "${PREFIX+x}" ]; then
+  if [ -n "${PREFIX}" ]; then
     rm -rf "$PREFIX"
   fi
 else
